KAN-486 refactor: Category 컴포넌트 분리
