+++ src/pages/home/index.jsx
@@ -1,14 +1,13 @@
-import React, { useState, useEffect } from "react";
-import ProductList from "../../components/Product/ProductList";
-import CategorySkeleton from "../../components/Skeletons/CategorySkeleton";
-import { Link } from "react-router-dom";
-import ProductSlider from "./ProductSlider";
+import { useState, useEffect } from "react";
 import { useProductsQuery } from "../../apis";
+
+import ProductList from "../../components/Product/ProductList";
+import ProductSlider from "./ProductSlider";
+import CategoryGrid from "./CategoryGrid";
 
 const Home = () => {
   const [categories, setCategories] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
-  const [expandedCategory, setExpandedCategory] = useState(null);
 
   useEffect(() => {
     const fetchCategories = async () => {
@@ -67,64 +66,10 @@ const Home = () => {
     fetchCategories();
   }, []);
 
-  const handleCategoryClick = categoryName => {
-    if (expandedCategory === categoryName) {
-      setExpandedCategory(null);
-    } else {
-      setExpandedCategory(categoryName);
-    }
-  };
-
-  const renderCategories = () => {
-    const rows = [];
-    for (let i = 0; i < categories.length; i += 3) {
-      const rowCategories = categories.slice(i, i + 3);
-      rows.push(
-        <React.Fragment key={i}>
-          <div className='mb-3 grid grid-cols-3 gap-4'>
-            {rowCategories.map(category => (
-              <div
-                key={category.name}
-                className='flex cursor-pointer flex-col items-center justify-center'
-                onClick={() => handleCategoryClick(category.name)}>
-                <div className='mb-2 text-3xl'>{category.icon}</div>
-                <span className='text-center text-sm'>{category.name}</span>
-              </div>
-            ))}
-          </div>
-          {expandedCategory && rowCategories.some(cat => cat.name === expandedCategory) && (
-            <div className='col-span-3 -mt-2 mb-3 rounded bg-gray-100 p-2'>
-              <div className='grid grid-cols-3 gap-2'>
-                {categories
-                  .find(cat => cat.name === expandedCategory)
-                  .subCategories.map(subCategory => (
-                    <div key={subCategory} className='py-1 text-center text-sm'>
-                      <Link to={"/detailcategory"}>{subCategory}</Link>
-                    </div>
-                  ))}
-              </div>
-            </div>
-          )}
-        </React.Fragment>
-      );
-    }
-    return rows;
-  };
-
   return (
     <div>
       <ProductSlider />
-      <div className='noScrollbar flex-grow overflow-y-auto p-2'>
-        {isLoading ? (
-          <div className='grid grid-cols-3 gap-4'>
-            {Array.from({ length: 9 }).map((_, index) => (
-              <CategorySkeleton key={index} />
-            ))}
-          </div>
-        ) : (
-          renderCategories()
-        )}
-      </div>
+      <CategoryGrid categories={categories} isLoading={isLoading} />
       <ProductList useQueryHook={useProductsQuery} gridCols={1} />
     </div>
   );
